refactor(web): refactored login and register pages
