+++ apps/web/src/components/Auth/Register/Form.tsx
@@ -1,70 +1,59 @@
-import { Form, Controller } from "@litespace/ui/Form";
-import { Button } from "@litespace/ui/Button";
-import { useToast } from "@litespace/ui/Toast";
-import { useFormatMessage } from "@litespace/ui/hooks/intl";
-import React, { useCallback, useEffect, useMemo } from "react";
-import { useForm } from "react-hook-form";
-import { Link, useNavigate, useParams } from "react-router-dom";
-import { IUser } from "@litespace/types";
-import { useRegisterUser } from "@litespace/headless/user";
-import { useUser } from "@litespace/headless/context/user";
-import { Typography } from "@litespace/ui/Typography";
-import {
-  useValidateEmail,
-  useValidatePassword,
-} from "@litespace/ui/hooks/validation";
-import { useGoogle } from "@/hooks/google";
-import Google from "@litespace/assets/Google";
-import { Landing, Web } from "@litespace/utils/routes";
-import { router } from "@/lib/routes";
-import { useOnError } from "@/hooks/error";
 import { VerifyEmail } from "@/components/Common/VerifyEmail";
-import { useRender } from "@/hooks/render";
+import { useOnError } from "@/hooks/error";
+import { useGoogle } from "@/hooks/google";
+import { useRender } from "@litespace/headless/common";
+import { router } from "@/lib/routes";
+import Google from "@litespace/assets/Google";
+import { useUser } from "@litespace/headless/context/user";
+import { useForm } from "@litespace/headless/form";
+import { useRegisterUser } from "@litespace/headless/user";
+import { IUser } from "@litespace/types";
+import { Button } from "@litespace/ui/Button";
+import { Form } from "@litespace/ui/Form";
+import { useFormatMessage } from "@litespace/ui/hooks/intl";
+import { useMakeValidators } from "@litespace/ui/hooks/validation";
+import { Input, Password } from "@litespace/ui/Input";
+import { validateEmail, validatePassword } from "@litespace/ui/lib/validate";
+import { useToast } from "@litespace/ui/Toast";
+import { Typography } from "@litespace/ui/Typography";
+import { Landing, Web } from "@litespace/utils/routes";
+import { useCallback, useEffect, useMemo } from "react";
+import { Link, useNavigate, useParams } from "react-router-dom";
 
-interface IForm {
+type Form = {
   email: string;
   password: string;
-  confirmedPassword: string;
-}
+  confirmPassword: string;
+};
 
 type Role = "student" | "tutor";
 
 const RegisterForm: React.FC = () => {
   const intl = useFormatMessage();
-  const user = useUser();
-  const navigate = useNavigate();
   const toast = useToast();
-  const params = useParams<{ role: Role }>();
+  const user = useUser();
 
   const verifyEmailDialog = useRender();
 
-  const { watch, handleSubmit, control, formState } = useForm<IForm>({
-    defaultValues: {
-      email: "",
-      password: "",
-      confirmedPassword: "",
-    },
-  });
-  const errors = formState.errors;
+  const navigate = useNavigate();
 
-  const validateEmail = useValidateEmail(true);
-  const validatePassword = useValidatePassword(true);
+  // ======== Google Registeration ============
+  const params = useParams<{ role: Role }>();
   const role = useMemo(() => {
     if (params.role === "student") return IUser.Role.Student;
     if (params.role === "tutor") return IUser.Role.Tutor;
     return null;
   }, [params.role]);
-  const google = useGoogle({
-    role: role || undefined,
-  });
-  const email = watch("email");
-  const password = watch("password");
-  const confirmedPassword = watch("confirmedPassword");
 
   useEffect(() => {
     if (!role) return navigate(Web.Root);
   }, [navigate, role]);
 
+  const google = useGoogle({
+    role: role || undefined,
+  });
+
+  // ========== manual registeration ============
   const onSuccess = useCallback(
     async ({ user: info, token }: IUser.RegisterApiResponse) => {
       user.set({ user: info, token });
@@ -85,69 +74,83 @@ const RegisterForm: React.FC = () => {
 
   const mutation = useRegisterUser({ onSuccess, onError });
 
-  const onSubmit = useMemo(
-    () =>
-      handleSubmit((payload: IForm) => {
-        if (!role) return;
-        return mutation.mutate({
-          email: payload.email,
-          password: payload.password,
-          role,
-        });
-      }),
-    [handleSubmit, mutation, role]
-  );
+  // ============= form ==============
+  const validators = useMakeValidators<Form>({
+    email: {
+      required: true,
+      validate: validateEmail,
+    },
+    password: {
+      required: true,
+      validate: validatePassword,
+    },
+    confirmPassword: {
+      required: true,
+      validate: (value, form) => {
+        if (value !== form.password) return "register.password-not-the-same";
+        return validatePassword(value);
+      },
+    },
+  });
+
+  const form = useForm<Form>({
+    defaults: {
+      email: "",
+      password: "",
+      confirmPassword: "",
+    },
+    validators,
+    onSubmit: (data) => {
+      if (!role) return;
+      return mutation.mutate({
+        email: data.email,
+        password: data.password,
+        role,
+      });
+    },
+  });
 
   return (
-    <Form onSubmit={onSubmit}>
+    <Form onSubmit={form.onSubmit}>
       <div className="flex flex-col gap-4 sm:gap-6">
         <div className="flex flex-col gap-2 sm:gap-4">
-          <Controller.Input
+          <Input
             id="email"
-            control={control}
             name="email"
-            value={email}
+            value={form.state.email}
+            onChange={(e) => form.set("email", e.target.value)}
             autoComplete="off"
             label={intl("labels.email")}
-            rules={{ validate: validateEmail }}
             placeholder={intl("labels.email.placeholder")}
             disabled={mutation.isPending || google.loading}
-            state={errors.email ? "error" : undefined}
-            helper={errors.email?.message}
+            state={form.errors.email ? "error" : undefined}
+            helper={form.errors.email}
             idleDir="rtl"
           />
 
-          <Controller.Password
+          <Password
             idleDir="rtl"
             id="password"
             name="password"
-            value={password}
-            control={control}
+            value={form.state.password}
+            onChange={(e) => form.set("password", e.target.value)}
             label={intl("labels.password")}
-            helper={errors.password?.message}
-            rules={{ validate: validatePassword }}
-            state={errors.password ? "error" : undefined}
+            helper={form.errors.password}
+            state={form.errors.password ? "error" : undefined}
             disabled={mutation.isPending || google.loading}
             placeholder={intl("register.password.placeholder")}
           />
 
-          <Controller.Password
-            control={control}
-            id="confirmedPassword"
-            name="confirmedPassword"
-            value={confirmedPassword}
-            rules={{
-              validate: (value) => {
-                if (value !== watch("password"))
-                  return intl("register.password-not-the-same");
-                return validatePassword(value);
-              },
-            }}
+          <Password
+            id="confirmPassword"
+            name="confirmPassword"
+            value={form.state.confirmPassword}
+            onChange={(e) => form.set("confirmPassword", e.target.value)}
             disabled={mutation.isPending || google.loading}
             label={intl("register.confirm-password")}
             placeholder={intl("register.confirm-password.placeholder")}
-            state={errors.confirmedPassword ? "error" : undefined}
-            helper={errors.confirmedPassword?.message}
+            state={form.errors.confirmPassword ? "error" : undefined}
+            helper={form.errors.confirmPassword}
             idleDir="rtl"
           />
