--- packages/ui/src/components/SidebarNav/index.ts
@@ -1,1 +1,0 @@
-export { default as SidebarNav } from "@/components/SidebarNav/SidebarNav";
