--- packages/ui/src/components/SidebarNav/types.ts
@@ -1,9 +1,0 @@
-import { Void } from "@litespace/types";
-
-export type NavOption = {
-  label: string;
-  route: string;
-  onClick?: Void;
-  loading?: boolean;
-  disabled?: boolean;
-};
