+++ packages/ui/src/hooks/index.ts
@@ -2,8 +2,7 @@ export { useWithdrawMethod } from "@/hooks/withdraw";
 export { useInvoiceStatus } from "@/hooks/invoice";
 export { useFormatMessage } from "@/hooks/intl";
 export { useTheme, Theme } from "@/hooks/theme";
-export { useReload, useRender } from "@/hooks/common";
-export { useClosableRef } from "@/hooks/dom";
+export { useReload } from "@/hooks/common";
 export { useKeys } from "@/hooks/keys";
 export {
   useSelectedRoom,
