+++ packages/ui/src/index.ts
@@ -39,7 +39,6 @@ export { ForgetPasswordDialog } from "@/components/ForgetPasswordDialog";
 export { Loading } from "@/components/Loading";
 export { RawHtml } from "@/components/RawHtml";
 export { Price } from "@/components/Price";
-export { SidebarNav } from "@/components/SidebarNav";
 export { OnlineStatus, UserOnlineStatus } from "@/components/OnlineStatus";
 export { messages, locales, type LocalMap, type LocalId } from "@/locales";
 export { IconField } from "@/components/IconField";
